project 13 / final / errors handlers

+++ controllers/cards.js
@@ -3,6 +3,7 @@ const Card = require('../models/Card');
 const ERROR_VALIDATION = 400;
 const ERROR_COMMON = 500;
 const ERROR_NOT_FOUND = 404;
+const LENGTH_OF_ID = 24;
 
 function getAllCards(req, res) {
   Card.find({})
@@ -31,21 +32,40 @@ function createCard(req, res) {
 }
 
 function deleteCardById(req, res) {
+  if (req.params.cardId.length < LENGTH_OF_ID) {
+    res.status(ERROR_VALIDATION).send({ message: 'Переданы некорректные данные' });
+    return;
+  }
   Card.findByIdAndRemove(req.params.cardId)
-    .then((card) => res.send({ data: card }))
+    .then((card) => {
+      if (!card) {
+        return Promise.reject(new Error());
+      }
+      return res.send({ data: card });
+    })
     .catch(() => res.status(ERROR_NOT_FOUND).send({ message: 'Карточка с указанным id не найдена' }));
 }
 
 function addLike(req, res) {
+  if (req.params.cardId.length < LENGTH_OF_ID) {
+    res.status(ERROR_VALIDATION).send({ message: 'Переданы некорректные данные' });
+    return;
+  }
   Card.findByIdAndUpdate(
     req.params.cardId,
     { $addToSet: { likes: req.user._id } },
     { new: true },
   )
-    .then((likes) => res.send({ data: likes }))
+    .then((likes) => {
+      if (!likes) {
+        return Promise.reject(new Error('CastError'));
+      }
+      return res.send({ data: likes });
+    })
     .catch((err) => {
-      if (err.name === 'CastError') {
+      if (err.name === 'CastError' || err.message === 'CastError') {
         res.status(ERROR_NOT_FOUND).send({ message: 'Передан несуществующий id карточки' });
+        return;
       } if (err.name === 'ValidationError') {
         res.status(ERROR_VALIDATION).send({ message: 'Переданы некорректные данные для постановки/снятии лайка' });
       } else {
@@ -55,15 +75,25 @@ function addLike(req, res) {
 }
 
 function deleteLike(req, res) {
+  if (req.params.cardId.length < LENGTH_OF_ID) {
+    res.status(ERROR_VALIDATION).send({ message: 'Переданы некорректные данные' });
+    return;
+  }
   Card.findByIdAndUpdate(
     req.params.cardId,
     { $pull: { likes: req.user._id } },
     { new: true },
   )
-    .then((likes) => res.send({ data: likes }))
+    .then((likes) => {
+      if (!likes) {
+        return Promise.reject(new Error('CastError'));
+      }
+      return res.send({ data: likes });
+    })
     .catch((err) => {
-      if (err.name === 'CastError') {
+      if (err.name === 'CastError' || err.message === 'CastError') {
         res.status(ERROR_NOT_FOUND).send({ message: 'Передан несуществующий id карточки' });
+        return;
       } if (err.name === 'ValidationError') {
         res.status(ERROR_VALIDATION).send({ message: 'Переданы некорректные данные для постановки/снятии лайка' });
       } else {
